Cover page when popups appear and add task popups
A div was added to the html file to cover the rest of the page so no
other actions on the page are reachable while the popup is displayed.
This works for both the task popup and folder popup.

A task popup was created with some input for the user to add new tasks.